Implementation of q_insert
Finished q_insert_haed and q_insert_tail
q_insert_head can insert a node with a string value and put it to head
of queue.
q_insert_end is implemented with the same method with q_insert_head,
only different is that we insert this node to end of queue.