Implementation of q_ascend and q_descend
The function check whether the queue is ascend or descend by iterate all nodes of queue list_for_each_entry_safe() macro is used to target the aim that we
want to check.
We start comparing the value from "safe" which it next to
the "entry". And use strcmp() to compare value of the nodes.